tests/wpt/web-platform-tests/custom-elements/state/tentative/ElementInternals-states.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
class TestElement extends HTMLElement {
  constructor() {
    super();
    this._internals = this.attachInternals();
  }

  get internals() {
    return this._internals;
  }
}
customElements.define("test-element", TestElement);

test(() => {
  let i = (new TestElement()).internals;

  assert_true(i.states instanceof CustomStateSet);
  assert_equals(i.states.size, 0);
  assert_false(i.states.has('foo'));
  assert_false(i.states.has('--foo'));
  assert_equals(i.states.toString(), '[object CustomStateSet]');
}, 'CustomStateSet behavior of ElementInternals.states: Initial state');

test(() => {
  let i = (new TestElement()).internals;
  assert_throws_js(TypeError, () => { i.states.supports('foo'); });
  assert_throws_dom('SyntaxError', () => { i.states.add(''); });
  assert_throws_dom('SyntaxError', () => { i.states.add('--a\tb'); });
}, 'CustomStateSet behavior of ElementInternals.states: Exceptions');

test(() => {
  let i = (new TestElement()).internals;
  i.states.add('--foo');
  i.states.add('--bar');
  i.states.add('--foo');
  assert_equals(i.states.size, 2);
  assert_true(i.states.has('--foo'));
  assert_true(i.states.has('--bar'));
  assert_array_equals([...i.states], ['--foo', '--bar']);
  i.states.delete('--foo');
  assert_array_equals([...i.states], ['--bar']);
  i.states.add('--foo');
  assert_array_equals([...i.states], ['--bar', '--foo']);
  i.states.delete('--bar');
  i.states.add('--baz');
  assert_array_equals([...i.states], ['--foo', '--baz']);
}, 'CustomStateSet behavior of ElementInternals.states: Modifications');
</script>
